crawler/requirements.txt
brotlipy==0.7.0
bs4==0.0.1
colorama==0.4.4
conda==4.12.0
conda-content-trust==0+unknown
conda-package-handling==1.8.1
django-rest-framework==0.1.0
faker==18.3.1
mysql-connector-python==8.0.32
mysqlclient==2.1.1
pip-chill==1.0.1
progressbar==2.5
progressbar2==4.2.0
pymysql==1.0.2
pyopenssl==22.0.0
pysocks==1.7.1
selenium==4.8.3
tqdm==4.63.0
